rewrote the function code to MarkdownPreview.js  #2644 #2662
